module: add support to canonicalize alternative spellings
The README.md originally specified that you could use "none" to set the
configuration fields for uploading. However, this was never actually
supported.

Update the configuration loading so that "none" will canonicalize to "no",
and "yes" will canonicalize to "all", so that users can use either value
in the configuration file.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>