only log when we actually upload, not before
Instead of logging when we call the Maybe-Upload functions, move the
logging inside of the functions. This ensures that we only log the times
where we actually decide to upload a file, rather than logging before we
do.

This is especially important for when configuration has disabled uploading
to gw2raidar or dps.report.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>