add Log-And-Write-Output and use it where appropriate
Instead of only logging data to the file, allow logging to the file *and*
console via the Log-And-Write-Output function.

This function is useful for putting data into the log file and onto the
screen, such as when an exception occurs, or to inform the user that
things are actually happening.

Make use of this new function in most places. The places which still use
Log-Output generally are for things that may help when decoding the log
file later, but aren't as useful to immediately be displayed.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>