try/catch the first execution of simpleArcParse
In some cases, simpleArcParse may not be able to run, such as if an
antivirus software interferes. Use a try/catch block, and if this catch is
called, indicate to the user that it might be due to antivirus
interference.

Also update the README.md to mention that simpleArcParse.exe may be
prevented from running by some antivirus software.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>