update example configuration files with more values
Recent commits have added new configuration options, but failed to
update the example configuration files in all cases. This makes the
options more difficult to discover.

Add an "everything" guild to the multipleguilds config file, and make
sure that all of the guild options are in each config file as
documentation.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>