add script which loads module and config file
The load-module.ps1 script is useful for starting a powershell session
that has the module file and config object already loaded. This is
primarily useful for debugging, but can sometimes be useful if you want
to manually load and publish bosses.

It can be made easy to use via a shortcut which invokes powershell with
the -noexit option and the script name. The shortcut will then launch
a powershell window, load the module and config object, and then start
a normal powershell console.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>