implement support for customizing discord mapping
Add two new configuration options to the guild object. First, the
"show_players" setting can be set to a string indicating what type of
mapping to do.

If it is set to "none", then the player list will be left empty.

If it is set to "discord_only", then only the mapped names will be used,
and any extra account info will be discarded.

If it is set to "accounts_only", then the player list will only show
account names, and will not map to the discord equivalents.

If it is set to "discord_if_possible", then account names will be
switched to their discord name if available. Otherwise the account name
will be used. This is the previous behavior, and is now default if
a guild does not specify the config option.

Additionally, add the "prefix_players_text" option, which can be set to
a string that will be prefixed to the players list. For example, you can
set this to the discord id for @here, (similar to player id pings), and
then set the "show_players" to none, in order to replace the players
list with an "@here" ping.

Add new test cases for the Get-Discord-Players function, now that it is
significantly more complex. This helps ensure that the behavior of the
function actually matches the expectations.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>