simpleArcParse: make duration calculation more robust
Some evtc log files are lacking a LOGEND event. If the encounter fails,
and there is no LOGEND event, then the duration will not be calculated,
which can cause format-encounters.ps1 to break.

upload-logs.ps1 has already been fixed to avoid saving the duration if
the simpleArcParse returns nothing. However, we can simply use the
timestamp on the last logged event instead.

Update the code in simpleArcParse to save this event time, and then use
the most accurate time we have when calculating the value of
precise_end.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>