refactor how we split bosses into segments
Modify the Split-Bosses function to just split by a single key, and
allow passing the key in. This simplifies the function, by removing the
now unnecessary Get-SubHash function.

Change the Format-And-Publish-All function to work with the new
Split-Bosses function, by calling it at each stage.

The primary reason for this change is a future enhancement which will
allow posting each boss individually instead of posting collected bosses
together. For this to work, we'll need to be able to handle each guild's
set of bosses separately.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>