remove gw2raidar support entirely
gw2raidar support is problematic because of the nature of their uploads
requiring a separate process to obtain the links. Obtaining these
permalinks is fragile, and does not work every time. Additionally, the
gw2raidar devs have not responded to multiple requests to improve their
API.

Historically their website has been slow, and I don't even enable
uploads to it anymore for my logs.

Additionally, future updates to the way that we format encounters are
going to make it more difficult to continue supporting gw2raidar,
especially as I do not actively use it and will probably introduce
regressions.

Drop support for gw2raidar entirely. A developer who cares can fork the
previous release and maintain the support themselves, if desired.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>