Add a new publish_encounters option for guilds
Add a new configuration option to allow publishing even the failed
encounters to a guild. Remove the built-in block in
format-encounters.ps1 so that the Format-And-Publish-Some function
receives all encounters.

Also add an optional parameter to Format-And-Publish-Some to allow
overriding this check when running the function manually via the
load-module.ps1 console.

Since we already support "successful" also enable "failed" to only
report failures. Additionally enable "none" as a shorthand method to
temporarily disable publishing encounters for a guild, which may be
useful when testing.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>