simpleArcParse: allow extracting local start and end time
We already support calculating the local start and end time for an
encounter. Add new "local_start_time" and "local_end_time" options to
simpleArcParse, so that we can extract the exact start and end times.

This data will be used by a future commit to measure the total length of
a given block of encounters.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>