improve logic for detecting "blocks" of bosses
We use the date to determine which blocks of bosses belong together.
This can be problematic if a user of the script happens to run bosses
over the Midnight time barrier. The published bosses will accidentally
be reported as separate blocks.

Fix this by instead using the start/end time of each boss, and grouping
encounters that have a maximum gap of no more than an hour. This is
a long enough time that we should generally treat a long run of bosses
as one "block" while still separating out multiple blocks if a user
happens to wait to upload multiple blocks at once.

We still separate first by guild, and by the type (raid or fractal) as
well.

Ultimately, this should help ensure that groups of encounters are
separated out, without arbitrarily breaking encounters run over midnight
into more than one post.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>