show estimate of the total duration of the encounters
Update the Format-And-Publish-Some function to estimate the total time
spent during a block of encounters, from the moment the first encounter
started, up to the moment the final encounter ended.

Add this to the headline, so that log viewers can get an estimate of how
long was spent for the total block of encounters.

Calculate this time *before* we strip the failed encounters, so that the
total time includes wipes, not just the time between the first and last
success.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>