fix display order of the wings of an encounter group
Previously we used the "time" property of a boss to sort the wings list.
However, this was recently changed to start_time, and the sort for this
was not updated. Additionally, we now sort the array of bosses up front,
so we don't need to re-sort the list anyways. Just remove the extraneous
sort, which corrects the display problem.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>