don't crash if no guild matches
Instead of crashing if a guild matches, modify the code to accept that
some encounters will not have a guild.

Update the Format-And-Publish-All function to not attempt publishing to
a $null guild. Also allow publishing those encounters to an "everything"
guild, if desired.

Update upload-logs.ps1 to avoid throwing an exception if no guild is
found. Instead, don't save the log.

Update Load-From-EVTC, so that it will accept a boss which has no guild
info. Additionally, modify the Format-And-Publish-Some so that if the
running_guild can't be found, we fall back to the publishing guild.

With these changes, it should now be possible to use the script even
when some encounters you run do not have associated guild information.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>