separate training golem encounters from fractals and raids
Use "encounter_type" to track whether an encounter was a golem,
a fractal, or a raid boss.

This way we can separate golem encounters from the other posts, and
avoid attempting to post "wings", when training golems all reside in the
special forces training area.

Also add the "golems" configuration value which controls whether we'll
post a golem encounter to a specific guild or not. This allows
configuring guilds to avoid having golem trial encounters published to
their guild discord.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>