simpleArcParse: convert output to JSON format
simpleArcParse began with a rather simple goal of extracting small bits
of useful data from EVTC files as fast as possible. To do this, each new
value was added as a separate keyword argument.

This has grown cumbersome for a couple of reasons. First, new output
must be added as a new keyword, and simpleArcParse must be called
multiple times, once for each keyword. Second, some kinds of data rely
on structure to convey information.

Replace the current method of extracting data with a simple JSON format.
This format can be extended easily. This avoids the need to run
simpleArcParse multiple times per file, meaning that we do not pay the
start up and search cost multiple times.

It also makes it rather simple to add character names and subgroup
information.

The new JSON format is a major change, so the version of simpleArcParse
is updated to v2.0.0. However, to help aid backwards compatibility,
the older keywords still remain for those users who rely on them
currently.

However, since it is expected that users will switch to the JSON format
going forward, simpleArcParse no longer attempts to figure out which
parts of the file to parse. It will always parse and load all the
details that it currently supports. This is done for simplicity in the
code, and is a known time trade off.

It will take some effort to refactor the upload and parsing scripts to
handle this new information format. This will be done in future patches.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>