update the structure and enumerations for EVTC files
The EVTC file format is documented on the ArcDPS website. The copies of
these structure definitions haven't been updated in the simpleArcParse
utility in a while.

Update the comments and structure definitions to match the latest
version on the website.

This should not include any new changes, except for the new state change
enumeration values. A few of the padding values in structures have been
renamed as they are now used, as well.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>