simpleArcParse: support extracting guild GUID information
Add a parse_guild_event handler to parse CBTS_GUILD events which
indicate the guild being represented by a character.

Store the 16byte value and perform the necessary byte-swaps to display
it as it is used by the GW2 API.

Extend the JSON format to add the "guid" value to a player character if
it has a valid guild id found.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>