re-write update-arcdps.ps1 to reduce duplicate code (#25)
* re-write update-arcdps.ps1 to reduce duplicate code

Refactor update-arcdps.ps1 to share the code for checking and
downloading the files.

Create an array of hash tables for each extension and the main ArcDPS
dll. Loop over the table to determine which files to update, then loop
over it again to actually update them.

Some of the files do not have their own hash file hosted on the web. To
handle these, loop to check for updates first. When updating each file,
also see if any other plugins should be updated at the same time.

This ensures that we always update extras and buildtemplates if the main
arcdps dll is updated.

Note that this version does not support downloading the experimental
ArcDPS dlls yet. Support for this may be restored in a future commit.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>

* add missing brackets for where clause

The use of .where requires both () and {} brackets.

Co-Authored-By: William Saulnier <lego6245@gmail.com>