upload-logs: don't stop immediately on upload failure
Instead of immediately exiting if a file fails to upload, modify its
LastWriteTime and continue. This ensures that we attempt to upload the
file again on a future run.

Additionally, keep track of all failed uploads and re-display them after
the upload finishes with a warning. This helps bring the failures to
attention in a more obvious way.

This ensures that we try to upload all encounters if possible, rather
than stopping early if one fails to parse.

If an encounter fails to parse, this can lead to continous attempts at
re-uploading it. A better solution in the future might be to handle
specific return errors from the dps.report API, and only bump the time
if dps.report was down. This way, encounters which fail to parse could
be skipped entirely.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>