refactor boss name mapping in simpleArcParse
simpleArcParse currently places data mapping boss id to name and
challenge mote status in separate places. This results in difficulty
when adding new boss IDs to the program.

Fix this by modifying the code to use a static map which will maintain
a structured set of data for each boss, instead of spreading the
information over multiple functions.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>