update the l0g-101086.psm1 to use location info
Modify the module file to expect location data from the EVTC information
generated by simpleArcParse.

This makes the scripts expect simpleArcParse version 2.4.0 at least.

With this change, there is a slight window where old encounters that
were recorded with a previous version of simpleArcParse will not be
published properly. However, newly published encounters will correctly
determine the location from the simpleArcParse data.

This has the advantage that new encounter data only needs to be added to
the simpleArcParse program, rather than needing  both a change to the
module file and to the program.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>