automatically add brackets in discord and emoji maps
Due to the way that JSON stores strings, the angle brackets in the emoji
map and discord player map must be escaped. This requires leaving
'\u003c' and '\u003e' escape sequences in the configuration file.

Allow the config file to forego these escape sequences by automatically
inserting brackets if they are missing.

This simplifies handling the addition of emoji icons and discord pings
to the configuration file.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>